Revert "Merge branch 'Character' of https://github.com/anhaeyeong/KOF_project into Develop"
This reverts commit 248460b5a47e1e05f966b01e2c6aff824f72c9d3, reversing
changes made to d448c0c97b9072c3ee1b192a8738987205a9fc25.

+++ 250304_WinAPI/Image.h
@@ -28,8 +28,8 @@ public:
 		// 애니메이션 관련
 		int maxFrameX;
 		int maxFrameY;
-		float frameWidth;
-		float frameHeight;
+		int frameWidth;
+		int frameHeight;
 		int currFrameX;
 		int currFrameY;
